Update html_resources.md
At the bottom of the tools sections I added a link for a website that allows the user to upload an image, then select a pixel within the image. It returns both a hexadecimal code as well as an RGB code that can be copied and pasted within code to easily select then implement specific colors.  I use this tool whenever I am making themes from an image as well as if I am being instructed to select a hyper specific color. 

I hope this is helpful to others!